Фикс рантайма

+++ code/modules/mob/living/carbon/human/human.dm
@@ -61,30 +61,31 @@
 		if(eta_status)
 			stat(null, eta_status)
 
-	if (client.statpanel == "Status")
+	if(client)
+		if (client.statpanel == "Status")
 
-		if (internal)
-			if (!internal.air_contents)
-				qdel(internal)
-			else
-				stat("Internal Atmosphere Info", internal.name)
-				stat("Tank Pressure", internal.air_contents.return_pressure())
-				stat("Distribution Pressure", internal.distribute_pressure)
+			if (internal)
+				if (!internal.air_contents)
+					qdel(internal)
+				else
+					stat("Internal Atmosphere Info", internal.name)
+					stat("Tank Pressure", internal.air_contents.return_pressure())
+					stat("Distribution Pressure", internal.distribute_pressure)
 
-		var/obj/item/organ/xenos/plasmavessel/P = internal_organs_by_name["plasma vessel"]
-		if(P)
-			stat(null, "Phoron Stored: [P.stored_plasma]/[P.max_plasma]")
+			var/obj/item/organ/xenos/plasmavessel/P = internal_organs_by_name["plasma vessel"]
+			if(P)
+				stat(null, "Phoron Stored: [P.stored_plasma]/[P.max_plasma]")
 
-		if(back && istype(back,/obj/item/weapon/rig))
-			var/obj/item/weapon/rig/suit = back
-			var/cell_status = "ERROR"
-			if(suit.cell) cell_status = "[suit.cell.charge]/[suit.cell.maxcharge]"
-			stat(null, "Suit charge: [cell_status]")
+			if(back && istype(back,/obj/item/weapon/rig))
+				var/obj/item/weapon/rig/suit = back
+				var/cell_status = "ERROR"
+				if(suit.cell) cell_status = "[suit.cell.charge]/[suit.cell.maxcharge]"
+				stat(null, "Suit charge: [cell_status]")
 
-		if(mind)
-			if(mind.changeling)
-				stat("Chemical Storage", mind.changeling.chem_charges)
-				stat("Genetic Damage Time", mind.changeling.geneticdamage)
+			if(mind)
+				if(mind.changeling)
+					stat("Chemical Storage", mind.changeling.chem_charges)
+					stat("Genetic Damage Time", mind.changeling.geneticdamage)
 
 /mob/living/carbon/human/ex_act(severity)
 	if(!blinded)
@@ -332,11 +333,11 @@
 		return get_id_name("Unknown")
 	if( head && (head.flags_inv&HIDEFACE) )
 		return get_id_name("Unknown")		//Likewise for hats
-	var/face_name = get_face_name()
-	var/id_name = get_id_name("")
+	//var/face_name = get_face_name()
+	/*var/id_name = get_id_name("")
 	if(id_name && (id_name != face_name))
-		return "[face_name] (as [id_name])"
-	return face_name
+		return "[face_name] (as [id_name])"*/
+	return get_face_name()
 
 //Returns "Unknown" if facially disfigured and real_name if not. Useful for setting name when polyacided or when updating a human's name variable
 /mob/living/carbon/human/proc/get_face_name()
